feat(css): Hide header and markdown border in print view
This adds CSS scoped to print media to hide the header and border around
the markdown body.

I love this markdown preview, and I use it all the time. Often, to draft
a document that I then print to PDF. Each time, I need to manually
override the style to remove the header and border when printing, so I
figured I'm perhaps not the only one so I would fix this.

This does only impact the style when printing.